update for reversing tooltips so it matches stacked chart
